Rename / remove animation classes
