+++ Source/Scene/ModelExperimental/ModelAnimationCollection.js
@@ -1,15 +1,15 @@
-import defaultValue from "../Core/defaultValue.js";
-import defined from "../Core/defined.js";
-import DeveloperError from "../Core/DeveloperError.js";
-import Event from "../Core/Event.js";
-import JulianDate from "../Core/JulianDate.js";
-import CesiumMath from "../Core/Math.js";
+import defaultValue from "../../Core/defaultValue.js";
+import defined from "../../Core/defined.js";
+import DeveloperError from "../../Core/DeveloperError.js";
+import Event from "../../Core/Event.js";
+import JulianDate from "../../Core/JulianDate.js";
+import CesiumMath from "../../Core/Math.js";
 import ModelAnimation from "./ModelAnimation.js";
-import ModelAnimationLoop from "./ModelAnimationLoop.js";
-import ModelAnimationState from "./ModelAnimationState.js";
+import ModelAnimationLoop from ".././ModelAnimationLoop.js";
+import ModelAnimationState from ".././ModelAnimationState.js";
 
 /**
- * A collection of active model animations.  Access this using {@link Model#activeAnimations}.
+ * A collection of active model animations. Access this using {@link Model#activeAnimations}.
  *
  * @alias ModelAnimationCollection
  * @internalConstructor
@@ -58,7 +58,7 @@ function ModelAnimationCollection(model) {
   this.animateWhilePaused = false;
 
   this._model = model;
-  this._scheduledAnimations = [];
+  this._runtimeAnimations = [];
   this._previousTime = undefined;
 }
 
@@ -73,19 +73,31 @@ Object.defineProperties(ModelAnimationCollection.prototype, {
    */
   length: {
     get: function () {
-      return this._scheduledAnimations.length;
+      return this._runtimeAnimations.length;
+    },
+  },
+
+  /**
+   * The model that owns this animation collection.
+   *
+   * @memberof ModelAnimationCollection.prototype
+   *
+   * @type {Model}
+   * @readonly
+   */
+  model: {
+    get: function () {
+      return this._model;
     },
   },
 });
 
-function add(collection, index, options) {
+function addAnimation(collection, animation, options) {
   const model = collection._model;
-  const animations = model._runtime.animations;
-  const animation = animations[index];
-  const scheduledAnimation = new ModelAnimation(options, model, animation);
-  collection._scheduledAnimations.push(scheduledAnimation);
-  collection.animationAdded.raiseEvent(model, scheduledAnimation);
-  return scheduledAnimation;
+  const runtimeAnimation = new ModelAnimation(model, animation, options);
+  collection._runtimeAnimations.push(runtimeAnimation);
+  collection.animationAdded.raiseEvent(model, runtimeAnimation);
+  return runtimeAnimation;
 }
 
 /**
@@ -98,9 +110,9 @@ function add(collection, index, options) {
  * @param {String} [options.name] The glTF animation name that identifies the animation. Must be defined if <code>options.index</code> is <code>undefined</code>.
  * @param {Number} [options.index] The glTF animation index that identifies the animation. Must be defined if <code>options.name</code> is <code>undefined</code>.
  * @param {JulianDate} [options.startTime] The scene time to start playing the animation.  When this is <code>undefined</code>, the animation starts at the next frame.
- * @param {Number} [options.delay=0.0] The delay, in seconds, from <code>startTime</code> to start playing.
+ * @param {Number} [options.delay=0.0] The delay, in seconds, from <code>startTime</code> to start playing. This will only affect the animation if <code>options.loop</code> is ModelAnimationLoop.NONE.
  * @param {JulianDate} [options.stopTime] The scene time to stop playing the animation.  When this is <code>undefined</code>, the animation is played for its full duration.
- * @param {Boolean} [options.removeOnStop=false] When <code>true</code>, the animation is removed after it stops playing.
+ * @param {Boolean} [options.removeOnStop=false] When <code>true</code>, the animation is removed after it stops playing. This will only affect the animation if <code>options.loop</code> is ModelAnimationLoop.NONE.
  * @param {Number} [options.multiplier=1.0] Values greater than <code>1.0</code> increase the speed that the animation is played relative to the scene clock speed; values less than <code>1.0</code> decrease the speed.
  * @param {Boolean} [options.reverse=false] When <code>true</code>, the animation is played in reverse.
  * @param {ModelAnimationLoop} [options.loop=ModelAnimationLoop.NONE] Determines if and how the animation is looped.
@@ -119,6 +131,7 @@ function add(collection, index, options) {
  *   name : 'animation name'
  * });
  *
+ * @example
  * // Example 2. Add an animation by index
  * model.activeAnimations.add({
  *   index : 0
@@ -131,11 +144,11 @@ function add(collection, index, options) {
  * const animation = model.activeAnimations.add({
  *   name : 'another animation name',
  *   startTime : startTime,
- *   delay : 0.0,                          // Play at startTime (default)
+ *   delay : 0.0,                                 // Play at startTime (default)
  *   stopTime : Cesium.JulianDate.addSeconds(startTime, 4.0, new Cesium.JulianDate()),
- *   removeOnStop : false,                 // Do not remove when animation stops (default)
- *   multiplier : 2.0,                        // Play at double speed
- *   reverse : true,                       // Play in reverse
+ *   removeOnStop : false,                        // Do not remove when animation stops (default)
+ *   multiplier : 2.0,                            // Play at double speed
+ *   reverse : true,                              // Play in reverse
  *   loop : Cesium.ModelAnimationLoop.REPEAT      // Loop the animation
  * });
  *
@@ -143,7 +156,7 @@ function add(collection, index, options) {
  *   console.log(`Animation started: ${animation.name}`);
  * });
  * animation.update.addEventListener(function(model, animation, time) {
- *   console.log(`Animation updated:  ${animation.name}. glTF animation time: ${time}`);
+ *   console.log(`Animation updated: ${animation.name}. glTF animation time: ${time}`);
  * });
  * animation.stop.addEventListener(function(model, animation) {
  *   console.log(`Animation stopped: ${animation.name}`);
@@ -153,14 +166,18 @@ ModelAnimationCollection.prototype.add = function (options) {
   options = defaultValue(options, defaultValue.EMPTY_OBJECT);
 
   const model = this._model;
-  const animations = model._runtime.animations;
 
   //>>includeStart('debug', pragmas.debug);
-  if (!defined(animations)) {
+  if (!model.ready) {
     throw new DeveloperError(
       "Animations are not loaded.  Wait for Model.readyPromise to resolve."
     );
   }
+  //>>includeEnd('debug');
+
+  const animations = model.sceneGraph.components.animations;
+
+  //>>includeStart('debug', pragmas.debug);
   if (!defined(options.name) && !defined(options.index)) {
     throw new DeveloperError(
       "Either options.name or options.index must be defined."
@@ -170,6 +187,7 @@ ModelAnimationCollection.prototype.add = function (options) {
   if (defined(options.multiplier) && options.multiplier <= 0.0) {
     throw new DeveloperError("options.multiplier must be greater than zero.");
   }
+
   if (
     defined(options.index) &&
     (options.index >= animations.length || options.index < 0)
@@ -178,12 +196,12 @@ ModelAnimationCollection.prototype.add = function (options) {
   }
   //>>includeEnd('debug');
 
-  if (defined(options.index)) {
-    return add(this, options.index, options);
+  let index = options.index;
+  if (defined(index)) {
+    return addAnimation(this, animations[index], options);
   }
 
   // Find the index of the animation with the given name
-  let index;
   const length = animations.length;
   for (let i = 0; i < length; ++i) {
     if (animations[i].name === options.name) {
@@ -198,41 +216,43 @@ ModelAnimationCollection.prototype.add = function (options) {
   }
   //>>includeEnd('debug');
 
-  return add(this, index, options);
+  return addAnimation(this, animations[index], options);
 };
 
 /**
- * Creates and adds an animation with the specified initial properties to the collection
- * for each animation in the model.
+ * Creates and adds animations with the specified initial properties to the collection
+ * for all animations in the model.
  * <p>
  * This raises the {@link ModelAnimationCollection#animationAdded} event for each model so, for example, a UI can stay in sync.
  * </p>
  *
  * @param {Object} [options] Object with the following properties:
- * @param {JulianDate} [options.startTime] The scene time to start playing the animations.  When this is <code>undefined</code>, the animations starts at the next frame.
- * @param {Number} [options.delay=0.0] The delay, in seconds, from <code>startTime</code> to start playing.
- * @param {JulianDate} [options.stopTime] The scene time to stop playing the animations.  When this is <code>undefined</code>, the animations are played for its full duration.
- * @param {Boolean} [options.removeOnStop=false] When <code>true</code>, the animations are removed after they stop playing.
+ * @param {JulianDate} [options.startTime] The scene time to start playing the animations. When this is <code>undefined</code>, the animations starts at the next frame.
+ * @param {Number} [options.delay=0.0] The delay, in seconds, from <code>startTime</code> to start playing. This will only affect the animation if <code>options.loop</code> is ModelAnimationLoop.NONE.
+ * @param {JulianDate} [options.stopTime] The scene time to stop playing the animations. When this is <code>undefined</code>, the animations are played for its full duration.
+ * @param {Boolean} [options.removeOnStop=false] When <code>true</code>, the animations are removed after they stop playing. This will only affect the animation if <code>options.loop</code> is ModelAnimationLoop.NONE.
  * @param {Number} [options.multiplier=1.0] Values greater than <code>1.0</code> increase the speed that the animations play relative to the scene clock speed; values less than <code>1.0</code> decrease the speed.
  * @param {Boolean} [options.reverse=false] When <code>true</code>, the animations are played in reverse.
  * @param {ModelAnimationLoop} [options.loop=ModelAnimationLoop.NONE] Determines if and how the animations are looped.
  * @param {ModelAnimation.AnimationTimeCallback} [options.animationTime=undefined] If defined, computes the local animation time for all of the animations.
  * @returns {ModelAnimation[]} An array of {@link ModelAnimation} objects, one for each animation added to the collection.  If there are no glTF animations, the array is empty.
  *
- * @exception {DeveloperError} Animations are not loaded.  Wait for the {@link Model#readyPromise} to resolve.
+ * @exception {DeveloperError} Animations are not loaded. Wait for the {@link Model#readyPromise} to resolve.
  * @exception {DeveloperError} options.multiplier must be greater than zero.
  *
  * @example
  * model.activeAnimations.addAll({
- *   multiplier : 0.5,                        // Play at half-speed
+ *   multiplier : 0.5,                            // Play at half-speed
  *   loop : Cesium.ModelAnimationLoop.REPEAT      // Loop the animations
  * });
  */
 ModelAnimationCollection.prototype.addAll = function (options) {
   options = defaultValue(options, defaultValue.EMPTY_OBJECT);
 
+  const model = this._model;
+
   //>>includeStart('debug', pragmas.debug);
-  if (!defined(this._model._runtime.animations)) {
+  if (!model.ready) {
     throw new DeveloperError(
       "Animations are not loaded.  Wait for Model.readyPromise to resolve."
     );
@@ -243,14 +263,15 @@ ModelAnimationCollection.prototype.addAll = function (options) {
   }
   //>>includeEnd('debug');
 
-  const scheduledAnimations = [];
-  const model = this._model;
-  const animations = model._runtime.animations;
+  const animations = model.sceneGraph.components.animations;
+
+  const addedAnimations = [];
   const length = animations.length;
   for (let i = 0; i < length; ++i) {
-    scheduledAnimations.push(add(this, i, options));
+    const animation = addAnimation(this, animations[i], options);
+    addedAnimations.push(animation);
   }
-  return scheduledAnimations;
+  return addedAnimations;
 };
 
 /**
@@ -259,11 +280,11 @@ ModelAnimationCollection.prototype.addAll = function (options) {
  * This raises the {@link ModelAnimationCollection#animationRemoved} event so, for example, a UI can stay in sync.
  * </p>
  * <p>
- * An animation can also be implicitly removed from the collection by setting {@link ModelAnimation#removeOnStop} to
+ * An animation can also be implicitly removed from the collection by setting {@link ModelAnimationCollection#removeOnStop} to
  * <code>true</code>.  The {@link ModelAnimationCollection#animationRemoved} event is still fired when the animation is removed.
  * </p>
  *
- * @param {ModelAnimation} animation The animation to remove.
+ * @param {ModelAnimation} runtimeAnimation The runtime animation to remove.
  * @returns {Boolean} <code>true</code> if the animation was removed; <code>false</code> if the animation was not found in the collection.
  *
  * @example
@@ -272,15 +293,17 @@ ModelAnimationCollection.prototype.addAll = function (options) {
  * });
  * model.activeAnimations.remove(a); // Returns true
  */
-ModelAnimationCollection.prototype.remove = function (animation) {
-  if (defined(animation)) {
-    const animations = this._scheduledAnimations;
-    const i = animations.indexOf(animation);
-    if (i !== -1) {
-      animations.splice(i, 1);
-      this.animationRemoved.raiseEvent(this._model, animation);
-      return true;
-    }
+ModelAnimationCollection.prototype.remove = function (runtimeAnimation) {
+  if (!defined(runtimeAnimation)) {
+    return false;
+  }
+
+  const animations = this._runtimeAnimations;
+  const i = animations.indexOf(runtimeAnimation);
+  if (i !== -1) {
+    animations.splice(i, 1);
+    this.animationRemoved.raiseEvent(this._model, runtimeAnimation);
+    return true;
   }
 
   return false;
@@ -295,10 +318,10 @@ ModelAnimationCollection.prototype.remove = function (animation) {
  */
 ModelAnimationCollection.prototype.removeAll = function () {
   const model = this._model;
-  const animations = this._scheduledAnimations;
+  const animations = this._runtimeAnimations;
   const length = animations.length;
 
-  this._scheduledAnimations = [];
+  this._runtimeAnimations.length = 0;
 
   for (let i = 0; i < length; ++i) {
     this.animationRemoved.raiseEvent(model, animations[i]);
@@ -308,12 +331,12 @@ ModelAnimationCollection.prototype.removeAll = function () {
 /**
  * Determines whether this collection contains a given animation.
  *
- * @param {ModelAnimation} animation The animation to check for.
+ * @param {ModelAnimation} runtimeAnimation The runtime animation to check for.
  * @returns {Boolean} <code>true</code> if this collection contains the animation, <code>false</code> otherwise.
  */
-ModelAnimationCollection.prototype.contains = function (animation) {
-  if (defined(animation)) {
-    return this._scheduledAnimations.indexOf(animation) !== -1;
+ModelAnimationCollection.prototype.contains = function (runtimeAnimation) {
+  if (defined(runtimeAnimation)) {
+    return this._runtimeAnimations.indexOf(runtimeAnimation) !== -1;
   }
 
   return false;
@@ -326,7 +349,7 @@ ModelAnimationCollection.prototype.contains = function (animation) {
  * all the animations in the collection.
  *
  * @param {Number} index The zero-based index of the animation.
- * @returns {ModelAnimation} The animation at the specified index.
+ * @returns {ModelAnimation} The runtime animation at the specified index.
  *
  * @example
  * // Output the names of all the animations in the collection.
@@ -341,18 +364,16 @@ ModelAnimationCollection.prototype.get = function (index) {
   if (!defined(index)) {
     throw new DeveloperError("index is required.");
   }
+
+  if (index >= this._runtimeAnimations.length || index < 0) {
+    throw new DeveloperError(
+      "index must be valid within the range of the collection"
+    );
+  }
   //>>includeEnd('debug');
 
-  return this._scheduledAnimations[index];
+  return this._runtimeAnimations[index];
 };
-
-function animateChannels(runtimeAnimation, localAnimationTime) {
-  const channelEvaluators = runtimeAnimation.channelEvaluators;
-  const length = channelEvaluators.length;
-  for (let i = 0; i < length; ++i) {
-    channelEvaluators[i](localAnimationTime);
-  }
-}
 
 const animationsToRemove = [];
 
@@ -367,14 +388,19 @@ function createAnimationRemovedFunction(
 }
 
 /**
+ * Updates the runtime animations in this collection, removing any animations
+ * that have stopped.
+ *
+ * @param {FrameState} frameState The current frame state.
+ * @returns {Boolean} <code>true</code> if an animation played during this update, <code>false</code> otherwise.
+ *
  * @private
  */
 ModelAnimationCollection.prototype.update = function (frameState) {
-  const scheduledAnimations = this._scheduledAnimations;
-  let length = scheduledAnimations.length;
+  const runtimeAnimations = this._runtimeAnimations;
+  let length = runtimeAnimations.length;
 
   if (length === 0) {
-    // No animations - quick return for performance
     this._previousTime = undefined;
     return false;
   }
@@ -387,44 +413,43 @@ ModelAnimationCollection.prototype.update = function (frameState) {
   }
   this._previousTime = JulianDate.clone(frameState.time, this._previousTime);
 
-  let animationOccured = false;
+  let animationOccurred = false;
   const sceneTime = frameState.time;
   const model = this._model;
 
   for (let i = 0; i < length; ++i) {
-    const scheduledAnimation = scheduledAnimations[i];
-    const runtimeAnimation = scheduledAnimation._runtimeAnimation;
+    const runtimeAnimation = runtimeAnimations[i];
 
-    if (!defined(scheduledAnimation._computedStartTime)) {
-      scheduledAnimation._computedStartTime = JulianDate.addSeconds(
-        defaultValue(scheduledAnimation.startTime, sceneTime),
-        scheduledAnimation.delay,
+    if (!defined(runtimeAnimation._computedStartTime)) {
+      runtimeAnimation._computedStartTime = JulianDate.addSeconds(
+        defaultValue(runtimeAnimation.startTime, sceneTime),
+        runtimeAnimation.delay,
         new JulianDate()
       );
     }
 
-    if (!defined(scheduledAnimation._duration)) {
-      scheduledAnimation._duration =
-        runtimeAnimation.stopTime * (1.0 / scheduledAnimation.multiplier);
+    if (!defined(runtimeAnimation._duration)) {
+      runtimeAnimation._duration =
+        runtimeAnimation.localStopTime * (1.0 / runtimeAnimation.multiplier);
     }
 
-    const startTime = scheduledAnimation._computedStartTime;
-    const duration = scheduledAnimation._duration;
-    const stopTime = scheduledAnimation.stopTime;
+    const startTime = runtimeAnimation._computedStartTime;
+    const duration = runtimeAnimation._duration;
+    const stopTime = runtimeAnimation.stopTime;
 
     const pastStartTime = JulianDate.lessThanOrEquals(startTime, sceneTime);
-    const pastStopTime =
+    const reachedStopTime =
       defined(stopTime) && JulianDate.greaterThan(sceneTime, stopTime);
 
     // [0.0, 1.0] normalized local animation time
     let delta = 0.0;
     if (duration !== 0.0) {
       const seconds = JulianDate.secondsDifference(
-        pastStopTime ? stopTime : sceneTime,
+        reachedStopTime ? stopTime : sceneTime,
         startTime
       );
-      delta = defined(scheduledAnimation._animationTime)
-        ? scheduledAnimation._animationTime(duration, seconds)
+      delta = defined(runtimeAnimation._animationTime)
+        ? runtimeAnimation._animationTime(duration, seconds)
         : seconds / duration;
     }
 
@@ -434,76 +459,74 @@ ModelAnimationCollection.prototype.update = function (frameState) {
     // * we did not reach a user-provided stop time.
 
     const repeat =
-      scheduledAnimation.loop === ModelAnimationLoop.REPEAT ||
-      scheduledAnimation.loop === ModelAnimationLoop.MIRRORED_REPEAT;
+      runtimeAnimation.loop === ModelAnimationLoop.REPEAT ||
+      runtimeAnimation.loop === ModelAnimationLoop.MIRRORED_REPEAT;
 
     const play =
-      (pastStartTime || (repeat && !defined(scheduledAnimation.startTime))) &&
+      (pastStartTime || (repeat && !defined(runtimeAnimation.startTime))) &&
       (delta <= 1.0 || repeat) &&
-      !pastStopTime;
+      !reachedStopTime;
 
-    if (delta === scheduledAnimation._prevAnimationDelta) {
+    if (delta === runtimeAnimation._prevAnimationDelta) {
       const animationStopped =
-        scheduledAnimation._state === ModelAnimationState.STOPPED;
+        runtimeAnimation._state === ModelAnimationState.STOPPED;
       // no change to delta, and no change to the animation state means we can
       // skip the update this time around.
       if (play !== animationStopped) {
         continue;
       }
     }
-    scheduledAnimation._prevAnimationDelta = delta;
+    runtimeAnimation._prevAnimationDelta = delta;
 
     // If it IS, or WAS, animating...
-    if (play || scheduledAnimation._state === ModelAnimationState.ANIMATING) {
-      // STOPPED -> ANIMATING state transition?
-      if (play && scheduledAnimation._state === ModelAnimationState.STOPPED) {
-        scheduledAnimation._state = ModelAnimationState.ANIMATING;
-        if (scheduledAnimation.start.numberOfListeners > 0) {
-          frameState.afterRender.push(scheduledAnimation._raiseStartEvent);
+    if (play || runtimeAnimation._state === ModelAnimationState.ANIMATING) {
+      // ...transition from STOPPED to ANIMATING
+      if (play && runtimeAnimation._state === ModelAnimationState.STOPPED) {
+        runtimeAnimation._state = ModelAnimationState.ANIMATING;
+        if (runtimeAnimation.start.numberOfListeners > 0) {
+          frameState.afterRender.push(runtimeAnimation._raiseStartEvent);
         }
       }
 
       // Truncate to [0.0, 1.0] for repeating animations
-      if (scheduledAnimation.loop === ModelAnimationLoop.REPEAT) {
+      if (runtimeAnimation.loop === ModelAnimationLoop.REPEAT) {
         delta = delta - Math.floor(delta);
-      } else if (
-        scheduledAnimation.loop === ModelAnimationLoop.MIRRORED_REPEAT
-      ) {
+      } else if (runtimeAnimation.loop === ModelAnimationLoop.MIRRORED_REPEAT) {
         const floor = Math.floor(delta);
         const fract = delta - floor;
         // When odd use (1.0 - fract) to mirror repeat
         delta = floor % 2 === 1.0 ? 1.0 - fract : fract;
       }
 
-      if (scheduledAnimation.reverse) {
+      if (runtimeAnimation.reverse) {
         delta = 1.0 - delta;
       }
 
-      let localAnimationTime = delta * duration * scheduledAnimation.multiplier;
+      let localAnimationTime = delta * duration * runtimeAnimation.multiplier;
       // Clamp in case floating-point roundoff goes outside the animation's first or last keyframe
       localAnimationTime = CesiumMath.clamp(
         localAnimationTime,
-        runtimeAnimation.startTime,
-        runtimeAnimation.stopTime
+        runtimeAnimation.localStartTime,
+        runtimeAnimation.localStopTime
       );
 
-      animateChannels(runtimeAnimation, localAnimationTime);
+      runtimeAnimation.animate(localAnimationTime);
 
-      if (scheduledAnimation.update.numberOfListeners > 0) {
-        scheduledAnimation._updateEventTime = localAnimationTime;
-        frameState.afterRender.push(scheduledAnimation._raiseUpdateEvent);
+      if (runtimeAnimation.update.numberOfListeners > 0) {
+        runtimeAnimation._updateEventTime = localAnimationTime;
+        frameState.afterRender.push(runtimeAnimation._raiseUpdateEvent);
       }
-      animationOccured = true;
+      animationOccurred = true;
 
       if (!play) {
-        // ANIMATING -> STOPPED state transition?
-        scheduledAnimation._state = ModelAnimationState.STOPPED;
-        if (scheduledAnimation.stop.numberOfListeners > 0) {
-          frameState.afterRender.push(scheduledAnimation._raiseStopEvent);
+        // transition from ANIMATING to STOPPED
+        runtimeAnimation._state = ModelAnimationState.STOPPED;
+        if (runtimeAnimation.stop.numberOfListeners > 0) {
+          frameState.afterRender.push(runtimeAnimation._raiseStopEvent);
         }
 
-        if (scheduledAnimation.removeOnStop) {
-          animationsToRemove.push(scheduledAnimation);
+        if (runtimeAnimation.removeOnStop) {
+          animationsToRemove.push(runtimeAnimation);
         }
       }
     }
@@ -513,16 +536,14 @@ ModelAnimationCollection.prototype.update = function (frameState) {
   length = animationsToRemove.length;
   for (let j = 0; j < length; ++j) {
     const animationToRemove = animationsToRemove[j];
-    scheduledAnimations.splice(
-      scheduledAnimations.indexOf(animationToRemove),
-      1
-    );
+    runtimeAnimations.splice(runtimeAnimations.indexOf(animationToRemove), 1);
     frameState.afterRender.push(
       createAnimationRemovedFunction(this, model, animationToRemove)
     );
   }
   animationsToRemove.length = 0;
 
-  return animationOccured;
+  return animationOccurred;
 };
+
 export default ModelAnimationCollection;
