+++ package.json
@@ -41,8 +41,8 @@
   },
   "type": "module",
   "dependencies": {
-    "@zip.js/zip.js": "^2.3.12",
     "@tweenjs/tween.js": "^18.6.4",
+    "@zip.js/zip.js": "^2.3.12",
     "autolinker": "^3.14.3",
     "bitmap-sdf": "^1.0.3",
     "dompurify": "^2.2.2",
